refactor: extract cli, channel, and provider submodules
