+++ src/channels/health.rs
@@ -1,0 +1,16 @@
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub(super) enum ChannelHealthState {
+    Healthy,
+    Unhealthy,
+    Timeout,
+}
+
+pub(super) fn classify_health_result(
+    result: &std::result::Result<bool, tokio::time::error::Elapsed>,
+) -> ChannelHealthState {
+    match result {
+        Ok(true) => ChannelHealthState::Healthy,
+        Ok(false) => ChannelHealthState::Unhealthy,
+        Err(_) => ChannelHealthState::Timeout,
+    }
+}
