+++ src/channels/registry.rs
@@ -1,0 +1,1 @@
+// Channel registry/config extraction placeholder for iterative refactors.
